refactor: rename `script_events.ts` to `ipc_listeners.ts`
