+++ packs/BP/scripts/index.ts
@@ -1,5 +1,5 @@
 import "./network_links/network_link_entity";
 import "./block_destroyed";
 import "./custom_components";
-import "./script_events";
+import "./ipc_listeners";
 import "./ui";
